requirements.txt
livekit-agents
livekit-plugins-openai
python-dotenv
livekit-plugins-silero
pydantic-settings>=2.7.1
livekit-api
flask[async]
flask
flask-cors
uvicorn
beanie
ollama